changed tests.py to web layer
