+++ docker-compose.yml
@@ -2,11 +2,9 @@ models:
    image: tp33/django
    external_links:
       - mysql:db
-   links:
-      - selenium_chrome:selenium_chrome
    volumes:
       - ./models/:/app
-   command: bash -c "pip install selenium && python ./manage.py makemigrations models && python ./manage.py migrate && python ./manage.py loaddata db.json && mod_wsgi-express start-server --reload-on-changes ./models/wsgi.py && docker run -it --name web -p 8001:8000"
+   command: bash -c "python ./manage.py makemigrations models && python ./manage.py migrate && python ./manage.py loaddata db.json && mod_wsgi-express start-server --reload-on-changes ./models/wsgi.py && docker run -it --name web -p 8001:8000"
 
 exp:
    image: tp33/django
@@ -24,11 +22,12 @@ web1:
    image: tp33/django
    links:
       - exp:exp-api
+      - selenium_chrome:selenium_chrome
    ports:
       - "8003:8000"
    volumes:
       - ./web/:/app
-   command: bash -c "pip install requests && mod_wsgi-express start-server --reload-on-changes ./web/wsgi.py"
+   command: bash -c "pip install selenium && pip install requests && mod_wsgi-express start-server --reload-on-changes ./web/wsgi.py"
 
 web2:
    image: tp33/django
@@ -72,6 +71,7 @@ load_balancer:
   links:
     - web1:web1
     - web2:web2
+    - selenium_chrome:selenium_chrome
   volumes:
     - ./hap/:/haproxy.cfg
